# docs/swagger.yaml
basePath: /
definitions:
...
  handler.LoginFactoryAdminRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  handler.RegisterFactoryAdminRequest:
    properties:
      email:
        type: string
      factory_id:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - factory_id
    - name
    - password
    type: object
...
host: localhost:8080
info:
  contact:
    email: Obie.kal22@gmail.com
    url: www.linkedin.com/in/obie-ananda-a87a64212
  description: API documentation for the FTGO PlashCash project.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: FTGO PlasCash Project
  version: "1.0"
paths:
...
  /factory-admin/login:
    post:
      consumes:
      - application/json
      description: Authenticates a factory admin using their email and password and
        returns a JWT token.
      parameters:
      - description: Login Factory Admin Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.LoginFactoryAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid email or password
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login for factory admin
      tags:
      - Factory Admin
  /factory-admin/register:
    post:
      consumes:
      - application/json
      description: Registers a new factory admin with a name, email, password, and
        associated factory ID.
      parameters:
      - description: Register Factory Admin Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterFactoryAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Admin registered successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or email already registered
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a factory admin
      tags:
      - Factory Admin
  /factory-admin/request/{request_id}:
    post:
      consumes:
      - application/json
      description: Processes a pending vendor request for a specific vending machine,
        updates vendor revenue, and resets the vending machine.
      parameters:
      - description: Request ID
        in: path
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Request processed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request ID or unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Request or vending machine not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Process a factory vendor request
      tags:
      - Factory Admin
...
swagger: "2.0"